Relocate apis.resources to container_service.clusters.open_apis
