--- bcs-app/backend/apis/resources/custom_object/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
