+++ bcs-app/backend/container_service/clusters/open_apis/deployment.py
@@ -13,9 +13,8 @@
 #
 from rest_framework.response import Response
 
+from backend.apis.views import BaseAPIViewSet
 from backend.resources.deployment import Deployment
-
-from ..views import BaseAPIViewSet
 
 
 class DeploymentViewSet(BaseAPIViewSet):
